// frontend/src/autogen/proto.d.ts
export interface IInitialize {
  /** Initialize userInfo */
  userInfo?: IUserInfo | null

  /** Initialize environmentInfo */
  environmentInfo?: IEnvironmentInfo | null

  /** Initialize sessionStatus */
  sessionStatus?: ISessionStatus | null

  /** Initialize commandLine */
  commandLine?: string | null

  /** Initialize sessionId */
  sessionId?: string | null
}

/** Represents an Initialize. */
export class Initialize implements IInitialize {
  /**
   * Constructs a new Initialize.
   * @param [properties] Properties to set
   */
  constructor(properties?: IInitialize)

  /** Initialize userInfo. */
  public userInfo?: IUserInfo | null

  /** Initialize environmentInfo. */
  public environmentInfo?: IEnvironmentInfo | null

  /** Initialize sessionStatus. */
  public sessionStatus?: ISessionStatus | null

  /** Initialize commandLine. */
  public commandLine: string

  /** Initialize sessionId. */
  public sessionId: string

  /**
   * Creates a new Initialize instance using the specified properties.
   * @param [properties] Properties to set
   * @returns Initialize instance
   */
  public static create(properties?: IInitialize): Initialize

  /**
   * Encodes the specified Initialize message. Does not implicitly {@link Initialize.verify|verify} messages.
   * @param message Initialize message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IInitialize,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified Initialize message, length delimited. Does not implicitly {@link Initialize.verify|verify} messages.
   * @param message Initialize message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IInitialize,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an Initialize message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns Initialize
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): Initialize

  /**
   * Decodes an Initialize message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns Initialize
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(
    reader: $protobuf.Reader | Uint8Array
  ): Initialize

  /**
   * Verifies an Initialize message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an Initialize message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns Initialize
   */
  public static fromObject(object: { [k: string]: any }): Initialize

  /**
   * Creates a plain object from an Initialize message. Also converts values to other types if specified.
   * @param message Initialize
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: Initialize,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this Initialize to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for Initialize
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}

/** Properties of a Config. */
export interface IConfig {
  /** Config gatherUsageStats */
  gatherUsageStats?: boolean | null

  /** Config maxCachedMessageAge */
  maxCachedMessageAge?: number | null

  /** Config mapboxToken */
  mapboxToken?: string | null

  /** Config allowRunOnSave */
  allowRunOnSave?: boolean | null

  /** Config hideTopBar */
  hideTopBar?: boolean | null

  /** Config hideSidebarNav */
  hideSidebarNav?: boolean | null
}

/** Represents a Config. */
export class Config implements IConfig {
  /**
   * Constructs a new Config.
   * @param [properties] Properties to set
   */
  constructor(properties?: IConfig)

  /** Config gatherUsageStats. */
  public gatherUsageStats: boolean

  /** Config maxCachedMessageAge. */
  public maxCachedMessageAge: number

  /** Config mapboxToken. */
  public mapboxToken: string

  /** Config allowRunOnSave. */
  public allowRunOnSave: boolean

  /** Config hideTopBar. */
  public hideTopBar: boolean

  /** Config hideSidebarNav. */
  public hideSidebarNav: boolean

  /**
   * Creates a new Config instance using the specified properties.
   * @param [properties] Properties to set
   * @returns Config instance
   */
  public static create(properties?: IConfig): Config

  /**
   * Encodes the specified Config message. Does not implicitly {@link Config.verify|verify} messages.
   * @param message Config message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IConfig,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified Config message, length delimited. Does not implicitly {@link Config.verify|verify} messages.
   * @param message Config message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IConfig,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes a Config message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns Config
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): Config

  /**
   * Decodes a Config message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns Config
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: $protobuf.Reader | Uint8Array): Config

  /**
   * Verifies a Config message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates a Config message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns Config
   */
  public static fromObject(object: { [k: string]: any }): Config

  /**
   * Creates a plain object from a Config message. Also converts values to other types if specified.
   * @param message Config
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: Config,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this Config to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for Config
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}
// ...
export interface IUserInfo {
  /** UserInfo installationId */
  installationId?: string | null

  /** UserInfo installationIdV3 */
  installationIdV3?: string | null

  /** UserInfo email */
  email?: string | null
}
// ...
export interface IEnvironmentInfo {
  /** EnvironmentInfo streamlitVersion */
  streamlitVersion?: string | null

  /** EnvironmentInfo pythonVersion */
  pythonVersion?: string | null
}

/** Represents an EnvironmentInfo. */
export class EnvironmentInfo implements IEnvironmentInfo {
  /**
   * Constructs a new EnvironmentInfo.
   * @param [properties] Properties to set
   */
  constructor(properties?: IEnvironmentInfo)

  /** EnvironmentInfo streamlitVersion. */
  public streamlitVersion: string

  /** EnvironmentInfo pythonVersion. */
  public pythonVersion: string

  /**
   * Creates a new EnvironmentInfo instance using the specified properties.
   * @param [properties] Properties to set
   * @returns EnvironmentInfo instance
   */
  public static create(properties?: IEnvironmentInfo): EnvironmentInfo

  /**
   * Encodes the specified EnvironmentInfo message. Does not implicitly {@link EnvironmentInfo.verify|verify} messages.
   * @param message EnvironmentInfo message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IEnvironmentInfo,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified EnvironmentInfo message, length delimited. Does not implicitly {@link EnvironmentInfo.verify|verify} messages.
   * @param message EnvironmentInfo message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IEnvironmentInfo,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an EnvironmentInfo message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns EnvironmentInfo
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): EnvironmentInfo

  /**
   * Decodes an EnvironmentInfo message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns EnvironmentInfo
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(
    reader: $protobuf.Reader | Uint8Array
  ): EnvironmentInfo

  /**
   * Verifies an EnvironmentInfo message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an EnvironmentInfo message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns EnvironmentInfo
   */
  public static fromObject(object: { [k: string]: any }): EnvironmentInfo

  /**
   * Creates a plain object from an EnvironmentInfo message. Also converts values to other types if specified.
   * @param message EnvironmentInfo
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: EnvironmentInfo,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this EnvironmentInfo to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for EnvironmentInfo
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}